refactor: storybook docs를 위해 타입을 한 파일에 작성
- Grid.type.ts 제거 > Grid에 props type 선언
- props 별 설명 추가
